Workaround for broken sort order when sortOrderType == 'score'
Changing sort order via the arrow buttons doesn't work when sortOrderType == 'score' or in other words the user has selected `Order by Relevance` which is also the default.
Workaround:
Disable the buttons for changing sort order when sorting by `Relevance`.

A better solution might be to actually enable sorting when sort order is set to `Relevance`. Or disallow it by design (because IMHO it can be argued that reversing a list that is ordered by relevance is never useful) and properly disable the arrow button.